LP-10 Fix abs warning when floats are used

+++ ground/openpilotgcs/src/plugins/dial/dialgadgetwidget.cpp
@@ -558,7 +558,7 @@ void DialGadgetWidget::rotateNeedles()
     int dialRun = 3;
     if (n2enabled) {
         double needle2Diff;
-        if (abs((needle2Value - needle2Target) * 10) > 5 && beSmooth) {
+        if (abs(int((needle2Value - needle2Target) * 10)) > 5 && beSmooth) {
             needle2Diff = (needle2Target - needle2Value) / 5;
         } else {
             needle2Diff = needle2Target - needle2Value;
@@ -590,7 +590,7 @@ void DialGadgetWidget::rotateNeedles()
 
     // We assume that needle1 always exists!
     double needle1Diff;
-    if ((abs((needle1Value - needle1Target) * 10) > 5) && beSmooth) {
+    if ((abs(int((needle1Value - needle1Target) * 10)) > 5) && beSmooth) {
         needle1Diff = (needle1Target - needle1Value) / 5;
     } else {
         needle1Diff = needle1Target - needle1Value;
@@ -614,7 +614,7 @@ void DialGadgetWidget::rotateNeedles()
 
     if (n3enabled) {
         double needle3Diff;
-        if ((abs((needle3Value - needle3Target) * 10) > 5) && beSmooth) {
+        if ((abs(int((needle3Value - needle3Target) * 10)) > 5) && beSmooth) {
             needle3Diff = (needle3Target - needle3Value) / 5;
         } else {
             needle3Diff = needle3Target - needle3Value;
